requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.21.1
streamlit==1.28.0
transformers==4.34.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pandas==2.1.0
scikit-learn==1.3.0
python-dotenv==1.0.1
requests==2.31.0
pytest==7.4.0
openai>=1.0.0
